requirements.txt
aiofiles==25.1.0
fastapi==0.131.0
Jinja2==3.1.6
python-dotenv==1.2.1
python-multipart==0.0.22
uvicorn==0.41.0
websockets==16.0